requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
bitsandbytes>=0.41.0
accelerate>=0.24.0 